Fix signedness warning in gio/tests/socket-common.c
gio/tests/socket-common.c: In function ‘socket_address_from_string’:
gio/tests/socket-common.c:50:17: error: comparison of integer expressions of different signedness: ‘int’ and ‘long unsigned int’ [-Werror=sign-compare]
   for (i = 0; i < G_N_ELEMENTS (unix_socket_address_types); i++)
                 ^

+++ gio/tests/socket-common.c
@@ -45,7 +45,7 @@ static GSocketAddress *
 socket_address_from_string (const char *name)
 {
 #ifdef G_OS_UNIX
-  int i, len;
+  gsize i, len;
 
   for (i = 0; i < G_N_ELEMENTS (unix_socket_address_types); i++)
     {
